+++ heron/package/src/python/common/__init__.py
@@ -6,3 +6,4 @@ from .package import (
 
 from .utils import *
 from .constants import *
+from .parser import *
