+++ heron/package/src/python/common/constants.py
@@ -1,15 +1,15 @@
-ROLE = "role"
 CLUSTER = "cluster"
-ENV = "env"
+COUNTER = "counter"
+DELETED = "deleted"
 DESC = "description"
+DEST_PATH = "dest_path"
+ENV = "env"
+EXTRA = "extra_fields"
+LATEST = "latest"
+LIVE = "live"
 PKG = "package"
 PKG_PATH = "pkg_path"
-LIVE = "live"
-LATEST = "latest"
-DELETED = "deleted"
-EXTRA = "extra_fields"
+RAW = "raw"
+ROLE = "role"
 STORE_PATH = "store_location"
 VERSION = "version"
-DEST_PATH = "dest_path"
-COUNTER = "counter"
-RAW = "is_raw"
